[jschaul review] separate swagger module.
